requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

dateparser==1.2.0
fastapi==0.115.4
motor==3.6.0
openai==1.51.0
pydantic==2.9.2
pymongo==4.9.2
python-dotenv==1.0.1
ray==2.39.0
routers==0.10.1
tqdm==4.67.0
uvicorn==0.32.0
python-jose
passlib
python-multipart
apscheduler
email-validator
termcolor==2.4.0